requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pyyaml==6.0.2
aiosqlite==0.20.0
jinja2==3.1.4

# Data Collection
httpx==0.27.0
beautifulsoup4==4.12.3
feedparser==6.0.11
scrapeling==1.1.1

# Analytics & NLP
textblob==0.18.0
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4

# Scheduling
apscheduler==3.10.4

# WebSocket
websockets==12.0
